// src/models/SystemIdentifierType.ts
/**
 * DO NOT EDIT!
 * This file was automatically generated.
 */

export enum SystemIdentifierType {
	AccountUserName = 'AccountUserName',
	EmailAddress = 'EmailAddress',
	LisSourcedId = 'LisSourcedId',
	LtiContextId = 'LtiContextId',
	LtiDeploymentId = 'LtiDeploymentId',
	LtiPlatformId = 'LtiPlatformId',
	LtiToolId = 'LtiToolId',
	LtiUserId = 'LtiUserId',
	OneRosterSourcedId = 'OneRosterSourcedId',
	Other = 'Other',
	SisSourcedId = 'SisSourcedId',
	SystemId = 'SystemId',
}